Replace obsolete automake macro
In the current up-to-date OS X 10.8 and MacPorts, autoconf v2.69
complains that "AM_CONFIG_HEADER" is obsolete, and fails.  The macro it
suggests, AC_CONFIG_HEADERS, works just fine as a drop-in replacement.

The Sleuth Kit had a matching problem.  I've tested the same change in
both projects.

To test the macro change, I compiled Afflib's current head and ran the
built affinfo without arguments, before and after changing the macro.
I ran this test in the following additional distros (selected for still
being supported), with their up-to-date version of autoconf:

 * Fedora 17 (autoconf v2.68)
 * Fedora 18 (autoconf v2.69)
 * FreeBSD 9.1 (autoconf v2.69)
 * Ubuntu 8.04.4 LTS (autoconf v2.61)
 * Ubuntu 10.04.4 LTS (autoconf v2.65)
 * Ubuntu 12.10 (autoconf v2.69)

(The thoroughness is from wariness of autotool behaviors - MacPort's is
the only v2.69 autoconf to complain so far with such vehemence, but it
won't be forever.)

Signed-off-by: Alex Nelson <ajnelson@cs.ucsc.edu>